requirements_dev.txt
pip==9.0.1
bumpversion==0.5.3
wheel==0.30.0
watchdog==0.8.3
flake8==3.5.0
tox==2.9.1
coverage==4.4.2
Sphinx==1.6.6
cryptography==2.1.4
PyYAML==3.12
pytest==3.4.0
pytest-runner==3.0
pytorch==0.1.2
torchvision==0.2.0

